moovefree_outdoor/requirements.txt
opencv-python>=4.8.0
google-generativeai>=0.8.0
Pillow>=10.0.0
SpeechRecognition>=3.10.0
pyttsx3>=2.90
pyaudio>=0.2.13
numpy>=1.24.0
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
requests>=2.31.0
